cleanup usings. merge EventSources. move LiveMetrics Processors to subdirectory
