+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/Customizations/Models/CollectionConfigurationError.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-
 namespace Azure.Monitor.OpenTelemetry.AspNetCore.Models
 {
     internal partial class CollectionConfigurationError
