+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/Customizations/Models/CollectionConfigurationInfo.cs
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-using System.Collections.Generic;
 
 namespace Azure.Monitor.OpenTelemetry.AspNetCore.Models
 {
