+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/Customizations/Models/DocumentIngress.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Diagnostics;
-
 namespace Azure.Monitor.OpenTelemetry.AspNetCore.Models
 {
     /// <summary> Additional properties used to calculate metrics. </summary>
