+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/LiveMetrics/Filtering/FilterInfoPredicateUtility.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using Azure.Monitor.OpenTelemetry.AspNetCore.Models;
 
 namespace Azure.Monitor.OpenTelemetry.AspNetCore.LiveMetrics.Filtering
