+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/ConnectionString/ConnectionStringParser.cs
@@ -9,7 +9,7 @@ using Azure.Monitor.OpenTelemetry.Exporter.Internals.Diagnostics;
 #elif LIVE_METRICS_EXPORTER
 using Azure.Monitor.OpenTelemetry.LiveMetrics.Internals.Diagnostics;
 #elif ASP_NET_CORE_DISTRO
-using Azure.Monitor.OpenTelemetry.AspNetCore.LiveMetrics.Diagnostics;
+using Azure.Monitor.OpenTelemetry.AspNetCore;
 #endif
 
 // This alias is necessary because it will otherwise try to default to "Microsoft.Azure.Core" which doesn't exist.
@@ -54,8 +54,10 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals.ConnectionString
             {
 #if AZURE_MONITOR_EXPORTER
                 AzureMonitorExporterEventSource.Log.FailedToParseConnectionString(ex);
-#elif (LIVE_METRICS_EXPORTER || ASP_NET_CORE_DISTRO)
+#elif LIVE_METRICS_EXPORTER
                 LiveMetricsExporterEventSource.Log.FailedToParseConnectionString(ex);
+#elif ASP_NET_CORE_DISTRO
+                AzureMonitorAspNetCoreEventSource.Log.FailedToParseConnectionString(ex);
 #endif
                 throw new InvalidOperationException("Connection String Error: " + ex.Message, ex);
             }
