+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/Platform/DefaultPlatform.cs
@@ -13,7 +13,7 @@ using Azure.Monitor.OpenTelemetry.Exporter.Internals.Diagnostics;
 #elif LIVE_METRICS_EXPORTER
 using Azure.Monitor.OpenTelemetry.LiveMetrics.Internals.Diagnostics;
 #elif ASP_NET_CORE_DISTRO
-using Azure.Monitor.OpenTelemetry.AspNetCore.LiveMetrics.Diagnostics;
+using Azure.Monitor.OpenTelemetry.AspNetCore;
 #endif
 
 namespace Azure.Monitor.OpenTelemetry.Exporter.Internals.Platform
@@ -32,8 +32,10 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals.Platform
             {
 #if AZURE_MONITOR_EXPORTER
                 AzureMonitorExporterEventSource.Log.FailedToReadEnvironmentVariables(ex);
-#elif (LIVE_METRICS_EXPORTER || ASP_NET_CORE_DISTRO)
+#elif LIVE_METRICS_EXPORTER
                 LiveMetricsExporterEventSource.Log.FailedToReadEnvironmentVariables(ex);
+#elif ASP_NET_CORE_DISTRO
+                AzureMonitorAspNetCoreEventSource.Log.FailedToReadEnvironmentVariables(ex);
 #endif
                 _environmentVariables = new Dictionary<string, object>();
             }
