+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/SdkVersionUtils.cs
@@ -12,8 +12,7 @@ using Azure.Monitor.OpenTelemetry.LiveMetrics;
 using Azure.Monitor.OpenTelemetry.LiveMetrics.Internals.Diagnostics;
 #elif ASP_NET_CORE_DISTRO
 using Azure.Monitor.OpenTelemetry.AspNetCore;
-using Azure.Monitor.OpenTelemetry.AspNetCore.LiveMetrics.Diagnostics;
-
+using Azure.Monitor.OpenTelemetry.AspNetCore.LiveMetrics;
 #endif
 using OpenTelemetry;
 
@@ -64,8 +63,10 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
                 {
 #if AZURE_MONITOR_EXPORTER
                     AzureMonitorExporterEventSource.Log.VersionStringUnexpectedLength(type.Name, versionString);
-#elif (LIVE_METRICS_EXPORTER || ASP_NET_CORE_DISTRO)
+#elif LIVE_METRICS_EXPORTER
                     LiveMetricsExporterEventSource.Log.VersionStringUnexpectedLength(type.Name, versionString);
+#elif ASP_NET_CORE_DISTRO
+                    AzureMonitorAspNetCoreEventSource.Log.VersionStringUnexpectedLength(type.Name, versionString);
 #endif
                     return shortVersion.Substring(0, 20);
                 }
@@ -76,8 +77,10 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
             {
 #if AZURE_MONITOR_EXPORTER
                 AzureMonitorExporterEventSource.Log.ErrorInitializingPartOfSdkVersion(type.Name, ex);
-#elif (LIVE_METRICS_EXPORTER || ASP_NET_CORE_DISTRO)
+#elif LIVE_METRICS_EXPORTER
                 LiveMetricsExporterEventSource.Log.ErrorInitializingPartOfSdkVersion(type.Name, ex);
+#elif ASP_NET_CORE_DISTRO
+                AzureMonitorAspNetCoreEventSource.Log.ErrorInitializingPartOfSdkVersion(type.Name, ex);
 #endif
                 return null;
             }
@@ -107,8 +110,10 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
             {
 #if AZURE_MONITOR_EXPORTER
                 AzureMonitorExporterEventSource.Log.SdkVersionCreateFailed(ex);
-#elif (LIVE_METRICS_EXPORTER || ASP_NET_CORE_DISTRO)
+#elif LIVE_METRICS_EXPORTER
                 LiveMetricsExporterEventSource.Log.SdkVersionCreateFailed(ex);
+#elif ASP_NET_CORE_DISTRO
+                AzureMonitorAspNetCoreEventSource.Log.SdkVersionCreateFailed(ex);
 #endif
                 return null;
             }
